+++ fast/stages/00-bootstrap/automation.tf
@@ -34,13 +34,13 @@ module "automation-project" {
   }
   # machine (service accounts) IAM bindings
   iam = {
-    "roles/owner" = [module.automation-tf-bootstrap-sa.iam_email]
+    "roles/owner" = [
+      module.automation-tf-bootstrap-sa.iam_email
+    ]
     "roles/iam.serviceAccountAdmin" = [
-      module.automation-tf-bootstrap-sa.iam_email,
       module.automation-tf-resman-sa.iam_email
     ]
     "roles/storage.admin" = [
-      module.automation-tf-bootstrap-sa.iam_email,
       module.automation-tf-resman-sa.iam_email
     ]
   }